+++ src/app/station-board/map/station-board-map.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { SbbDialog } from '@sbb-esta/angular/dialog';
 
 import mapboxgl from 'mapbox-gl'
+import * as OJP from 'ojp-sdk'
 
 import { MapDebugControl } from 'src/app/map/controls/map-debug-control';
 import { MapLayersLegendControl } from 'src/app/map/controls/map-layers-legend-control';
@@ -9,9 +10,9 @@ import { MapService } from 'src/app/shared/services/map.service';
 import { UserTripService } from 'src/app/shared/services/user-trip.service';
 import { StationBoardService } from '../station-board.service';
 
-import * as OJP from 'ojp-sdk'
 import { StopEventServiceRenderer } from './stop-event-service-renderer/stop-event-service-renderer';
 import { MapHelpers } from 'src/app/map/helpers/map.helpers';
+import { LanguageService } from 'src/app/shared/services/language.service';
 
 @Component({
   selector: 'station-board-map',
@@ -26,7 +27,8 @@ export class StationBoardMapComponent implements OnInit {
     private userTripService: UserTripService,
     private mapService: MapService,
     private debugXmlPopover: SbbDialog,
-    private stationBoardService: StationBoardService
+    private stationBoardService: StationBoardService,
+    private languageService: LanguageService
   ) {
     this.mapLoadingPromise = null;
     this.stopEventServiceRenderer = null;
@@ -84,7 +86,7 @@ export class StationBoardMapComponent implements OnInit {
     const debugControl = new MapDebugControl(map);
     map.addControl(debugControl, 'top-left');
 
-    const mapLayersLegendControl = new MapLayersLegendControl(map, this.debugXmlPopover, this.userTripService);
+    const mapLayersLegendControl = new MapLayersLegendControl(map, this.debugXmlPopover, this.userTripService, this.languageService);
     map.addControl(mapLayersLegendControl, 'top-right');
   }